+++ modules/graphics/src/include/GOWindow.hpp
@@ -47,6 +47,7 @@ protected:
     unsigned m_id;
     GOFigure* m_goFig;
     int m_clickX, m_clickY;
+    std::wstring graphicsRootPath;
 
 public:
     GOWindow(unsigned id);